mcp_server/requirements.txt
fastapi
uvicorn[standard]
gunicorn
langgraph>=0.2.6
python-dotenv>=1.0.1
langgraph-swarm>=0.0.11
boto3
pinecone
joblib>=1.0.0,<2.0.0    
scikit-learn>=1.0.0,<2.0.0
catboost>=1.0.0,<2.0.0
imbalanced-learn == 0.13.0
pandas>=2.0.0,<3.0.0
matplotlib>=3.0.0,<4.0.0
pillow>=9.0.0,<10.0.0
langchain-community>=0.0.11
langgraph-cli>=0.2.8
langgraph-cli[inmem]
openai
langchain_openai
langgraph-checkpoint-postgres
psycopg[binary,pool]
psycopg2-binary
langchain-pinecone
langchain-experimental
PyGithub
langchain-aws
pinecone[grpc]
fastmcp